add search helpers file
